harden vehicle ownership, soft delete and validation

+++ backend/src/main/java/com/hofnarrxx/autolog/dto/PublicVehicleAccessResponse.java
@@ -7,7 +7,7 @@ public record PublicVehicleAccessResponse(
         String brand,
         String model,
         String fuelType,
-        Double mileage,
+        Integer mileage,
         Integer year,
         FuelSummaryResponse fuelSummary,
         MaintenanceSummaryResponse maintenanceSummary
